requirements.txt
requests
beautifulsoup4
python-dotenv
colorama
python-telegram-bot
slack_sdk
Flask
flask-cors
selenium
playwright 